refund using explicit lock
(not needed, IMHO)

in the way these financial transactions are designed, I highly
believe that theres no need for a lock, only a check after it
is enough...

but I did it anyway, just for the sake of documenting how
to do it in Ecto, only for the refund function.

now, if anyone tries to begin the same transaction, it will have to
wait, but reading operations even in transactions are still cool.